Refactor SidebarMenu component to include a Settings option with a theme toggle
The SidebarMenu component has been refactored to include a Settings option in the sidebar. This change adds a new SidebarItem with the name "Settings" and a gear icon. Clicking on this item will navigate the user to the "/settings" path. Additionally, a ThemeToggle component has been added to the bottom of the sidebar.