Add ListItemContainer component
This commit adds the ListItemContainer component, which is responsible for rendering a list of items. The component takes in an array of items and a renderItem function to define how each item should be rendered. This component will be used to encapsulate the logic for rendering the list items in a separate container component, allowing for better code organization and reusability.