feat: Add local database support with LocalForage
This commit adds support for a local database using the LocalForage library. A new file `db.ts` has been created, which defines a `Database` class with methods for getting, setting, removing, and clearing data. In addition, a `Storage` class has been created that creates an instance of the `Database` class. The local database can now be accessed through the `storage` object.

The inclusion of a local database will enable efficient storage and retrieval of data within the application, improving user experience and providing offline functionality.